+++ orocos/HawkEye/src/Obstacle.hpp
@@ -15,22 +15,22 @@ private:
 	std::vector<int>    _pos;
 	std::vector<double> _vel;
 	const shape_t _shape;
-	int _area;
+	double _area;
 
 public:
 	
-	Obstacle(int x, int y, double vx, double vy, shape_t shape, int area); //specific constructor
+	Obstacle(int x, int y, double vx, double vy, shape_t shape, double area); //specific constructor
 	virtual ~Obstacle(){} //virtual destructor to be able to delete inherited classes
 
 	void setPos(int x, int y);
 	void setVel(double vx, double vy);
 	// void setShape(shape_t shape);
-	void setArea(int area);
+	void setArea(double area);
 
 	std::vector<int> getPos();
 	std::vector<double> getVel();
 	shape_t getShape();
-	int getArea();
+	double getArea();
 
 protected:
 	static void obj2vec(Obstacle obstacle, std::vector<double> *obstacleVector);
